Prepare jest
Fix tests

Fix tests

Fix tests
