+++ jest.config.js
@@ -1,0 +1,18 @@
+export default {
+  moduleFileExtensions: ['ts', 'js'],
+  transform: {
+    '^.+\\.jsx?$': '<rootDir>/node_modules/ts-jest',
+    '^.+\\.tsx?$': '<rootDir>/node_modules/ts-jest'
+  },
+  testMatch: ['<rootDir>/src/**/?(*.)(spec|test).(j|t)s?(x)'],
+  collectCoverageFrom: ['src/**/*.{tsx,jsx}'],
+  testEnvironment: 'jest-environment-jsdom',
+  transformIgnorePatterns: [
+    '<rootDir>/node_modules/(?!(ol|antd|@babel|jest-runtime|(rc-*[a-z]*)|@ant-design|@terrestris))'
+  ],
+  moduleNameMapper: {
+    '\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$':
+      '<rootDir>/jest/fileMock.js',
+    '^.+\\.(css|less)$': '<rootDir>/jest/cssTransform.js'
+  }
+};
