+++ src/geoserver-publish-geotiff/index.js
@@ -77,3 +77,5 @@ const isGeoServerAvailable = async () => {
 
 // Initialize and start the worker process
 initialize(rabbitHost, rabbitUser, rabbitPass, workerQueue, resultQueue, geoserverPublishGeoTiff, isGeoServerAvailable, 10);
+
+export default geoserverPublishGeoTiff;
